move science soft-drivers to science
