+++ drivers/CMakeLists.txt
@@ -30,8 +30,6 @@ SOURCES
     src/color_sensor_opt4048.cpp
     src/mq4.cpp
     src/pressure_sensor_bme680.cpp
-    src/pump_manager.cpp
-    src/revolver.cpp
     src/scd40.cpp
     src/soil_sensor_sht21.cpp
     src/sk9822.cpp
